Move build tools to editor assembly
