+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildAndDeploy/BuildInfoExtensions.cs.meta
