+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildAndDeploy/IBuildInfo.cs
@@ -4,7 +4,7 @@ using UnityEditor;
 using UnityEditor.Build.Reporting;
 using UnityEngine;
 
-namespace XRTK.Utilities.Build
+namespace XRTK.Editor.BuildAndDeploy
 {
     /// <summary>
     /// The Build Info defines common properties for a build.
