+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildAndDeploy/MixedRealityAppIcon.cs.meta
