+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildAndDeploy/UnityPlayerBuildTools.cs
@@ -16,7 +16,7 @@ using XRTK.Extensions;
 using XRTK.Utilities.Editor;
 using Debug = UnityEngine.Debug;
 
-namespace XRTK.Utilities.Build
+namespace XRTK.Editor.BuildAndDeploy
 {
     /// <summary>
     /// Cross platform player build tools
