--- XRTK-Core/Packages/com.xrtk.core/Runtime/Utilities/BuildAndDeploy/XRTK.Utilities.Build.asmdef
@@ -1,18 +1,0 @@
-{
-    "name": "XRTK.Utilities.Build",
-    "references": [
-        "GUID:f3241d040533491e8a1e2714b27c3111",
-        "GUID:395bd8dd25074437ba735310022f8f0a"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": []
-}
