--- XRTK-Core/Packages/com.xrtk.core/Runtime/Utilities/BuildAndDeploy/XRTK.Utilities.Build.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b6de896d71a648ec9fa913e154ba8b46
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
